Clarify REQ 'timeout' field in TCP spec
We were a bit confused about what the behavior of the 'timeout'
parameter was for the TCP REQ command. We ended up having to dig into
the [source code](https://github.com/nsqio/nsq/blob/master/nsqd/channel.go#L368) to understand
what was happening, and thought we'd just update the docs with what
we found in the code.